[Refactoring] The root `vars.tf` are now only a reference to each module
In this Object-Oriented approach there's no reason to define the objects
two times (one on root and the other one on module). So we're removing
the hard-coded definition on root and allowing the variables to be
anything. The object passed to each variable are still matched to the
definition on module, so we still garantee the object follow the desired format.

This decreases the amount of duplicated code and facilitates the development.