Moved all vault related code to a sub-directory
We're going to initialize Vault manually for security reasons
(so the recovery and root keys doesn't get stored in any pipeline or state files).
In order to do that we need to separate the code of deploying vault to
the code of deploying the app. That way our final workflow will be:

1. Deploy Vault
2. Initialize Vault
3. Deploy Application